+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 
 ### 1.0.0 (2025-04-23)
 
-*   **Initial release** of the `version_checker` package.
+*   **Initial release** of the `check_app_version` package.
 
 *   Allows comparison of the current app version with the versions available in the Apple App Store (iOS) and Google Play Store (Android).